Move config to package.json. Closes #70
