--- config/index.json
@@ -1,7 +1,0 @@
-{
-  "destinationFolder": "dist",
-  "exportFileName": "library-dist",
-  "entryFileName": "index",
-  "exportVarName": "MyLibrary",
-  "mochaGlobals": ["stub", "spy", "expect"]
-}
