+++ package.json
@@ -52,5 +52,12 @@
     "sinon": "^1.12.2",
     "sinon-chai": "^2.6.0",
     "vinyl-source-stream": "^1.0.0"
+  },
+  "to5BoilerplateOptions": {
+    "destinationFolder": "dist",
+    "exportFileName": "library-dist",
+    "entryFileName": "index",
+    "exportVarName": "MyLibrary",
+    "mochaGlobals": ["stub", "spy", "expect"]
   }
 }
